% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_na_variable.R
\name{add_na_variable}
\alias{add_na_variable}
\title{Add NA Variable}
\usage{
add_na_variable(data = contacts_unflat, variable)
}
\arguments{
\item{data}{A data frame. Default is \code{contacts_unflat}.}

\item{variable}{A character vector specifying the variable(s) to add NA values to.}
}
\value{
The modified data frame with NA values added.
}
\description{
Add NA values to the specified variables in the data frame if they do not exist.
}
\examples{
add_na_variable(data = mtcars, variable = c("cyl", "var2"))

}
