% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_table.R
\name{bar_table}
\alias{bar_table}
\title{Creating box to be used in \code{PLH_shiny} function}
\usage{
bar_table(data, variable)
}
\value{
Box for use in \code{Shiny}
}
\description{
Creating box to be used in \code{PLH_shiny} function
}
