% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_function.R
\name{box_function}
\alias{box_function}
\title{Creating box to be used in \code{PLH_shiny} function}
\usage{
box_function(data_frame, text, colour, label_table, label_plot, variable)
}
\arguments{
\item{data_frame}{Data frame that contains the data to analyse.}

\item{text}{Variable containing titles for each box.}

\item{colour}{Variable containing colours for each box.}

\item{label_table}{Variable containing the table label for each box.}

\item{label_plot}{Variable containing the plot label for each box.}

\item{variable}{Variable containing the variable for each box.}
}
\value{
Box for use in \code{Shiny}
}
\description{
Creating box to be used in \code{PLH_shiny} function
}
