% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variables_existence.R
\name{check_variables_existence}
\alias{check_variables_existence}
\title{Check Variables Existence}
\usage{
check_variables_existence(df, data_frame)
}
\arguments{
\item{df}{A data frame containing the variable names and data frame names.}

\item{data_frame}{The default data frame to check against if data frame names are not provided in df.}
}
\value{
A data frame containing the results of variable existence checks.
}
\description{
Check the existence of variables in the specified data frames.
}
\examples{
#check_variables_existence(df = mtcars, data_frame = "my_default_df")

}
