% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_from_ODK_functions.R
\name{create_from_type_ODK}
\alias{create_from_type_ODK}
\title{Generate Data Frame of Box Items from Metadata}
\usage{
create_from_type_ODK(
  metadata = ODK_sample_survey_group,
  value = "boxplot_summary"
)
}
\arguments{
\item{metadata}{Data frame. Metadata to be processed, containing columns \code{type}, \code{label}, and \code{name}.}

\item{value}{Character. Default value for box items, typically used for "boxplot_summary" or other summary types.}
}
\value{
A data frame containing rows of box items, each representing metadata attributes.
}
\description{
\code{create_from_type_ODK} processes metadata to create a data frame with "box" items for each row in the metadata, based on the type of data.
}
