% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tab_items.R
\name{create_tab_items}
\alias{create_tab_items}
\title{Title}
\usage{
create_tab_items(data_list, d_box, status, colour)
}
\arguments{
\item{data_list}{todo}

\item{d_box}{todo}

\item{status}{todo}

\item{colour}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
 #todo
}
