% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_contents.R
\name{display_contents}
\alias{display_contents}
\title{Title TODO}
\usage{
display_contents(
  contents1 = contents,
  data_frame,
  data_list,
  loop = NULL,
  k = 1
)
}
\arguments{
\item{contents1}{TODO}

\item{data_frame}{TODO}

\item{data_list}{TODO}

\item{loop}{TODO}

\item{k}{TODO}
}
\value{
TODO
}
\description{
Title TODO
}
\examples{
#todo
}
