% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_from_ODK_functions.R
\name{generate_data_list}
\alias{generate_data_list}
\title{Generate Data List for Display in Excel Template}
\usage{
generate_data_list(input_data, output_data)
}
\arguments{
\item{input_data}{Data frame containing metadata of form variables, including their \code{name}, \code{type}, and \code{label}.}

\item{output_data}{Data frame containing the dataset, which includes variables to be checked against \code{input_data}.}
}
\value{
A list named \code{excel_template} structured as follows:
\itemize{
\item \code{contents}: A data frame with \code{name}, \code{type}, \code{ID}, and \code{icon} for each displayable group.
\item \code{main_page}: A hardcoded main page data frame.
\item Group-specific display tabs generated by \code{create_display_pages_from_metadata}.
}
}
\description{
This function generates a data list structured as an Excel template, with grouped and miscellaneous variables displayed
for creating tabs on a Shiny dashboard. It filters variables based on presence in both the form and dataset,
organises them into specified groups, and adds miscellaneous items as a separate tab if needed.
}
\details{
This function filters variables that are defined in the form but not present in the dataset and organises
the contents into groups for tabbed display. Variables without a specific group are organised under "miscellaneous."
The resulting list includes \code{contents}, a \code{main_page}, and individual display tabs for each group.
}
