% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_download.R
\name{get_data_to_download}
\alias{get_data_to_download}
\title{Title}
\usage{
get_data_to_download(data_to_download, i)
}
\arguments{
\item{data_to_download}{todo}

\item{i}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
 #todo
}
