% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metabase_data.R
\name{get_metabase_data}
\alias{get_metabase_data}
\title{Get data from metabase site}
\usage{
get_metabase_data(site, name = "app_users")
}
\arguments{
\item{site}{Connection to database to get original data (using DBI::dbConnect).}

\item{name}{Data to call from connection. Default \code{"app_users"}, but also takes \code{"app_notification_interaction"}.}
}
\value{
Data from Metabase site
}
\description{
Call ParentApp data from Metabase
}
\examples{
# TODO
}
