% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameter_value.R
\name{get_parameter_value}
\alias{get_parameter_value}
\title{Get Parameter Value}
\usage{
get_parameter_value(spreadsheet_parameters, name = "label", list = FALSE)
}
\arguments{
\item{spreadsheet_parameters}{A character string containing the spreadsheet parameters.}

\item{name}{The name of the parameter whose value you want to extract. Default is "label".}

\item{list}{Logical value indicating whether the parameter value should be treated as a list.}
}
\value{
The value of the specified parameter. If the parameter is not found, it returns NULL.
}
\description{
This function extracts a parameter value from a string that represents a set of spreadsheet parameters.
}
\examples{
# Extract a parameter value as a string
param_string <- 'label = "Value"'
get_parameter_value(param_string)  # Returns "Value"

# Extract a parameter value as a list
param_list <- 'options = c("Option1", "Option2", "Option3")'
get_parameter_value(param_list, "options", list = TRUE)  # Returns a character vector c("Option1", "Option2", "Option3")

# Parameter not found
param_not_found <- 'other_param = 42'
get_parameter_value(param_not_found)
}
