% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_assert.R
\name{hasCssClass}
\alias{hasCssClass}
\title{Check if a Shiny tag has a specified CSS class}
\usage{
hasCssClass(tag, class)
}
\arguments{
\item{tag}{A \code{shiny.tag} object (e.g., a \code{div} or \code{span} created with Shiny functions).}

\item{class}{A character string specifying the CSS class to check for.}
}
\value{
Returns \code{TRUE} if the \code{shiny.tag} object has the specified CSS class, \code{FALSE} otherwise.
}
\description{
This function is a version of the \code{hasCssClass} function
from the \code{shinydashboard} package. It checks whether a
\code{shiny.tag} object contains a specific CSS class.
}
\details{
This function is from the internal \code{hasCssClass} function
in the \code{shinydashboard} package, which performs similar checks on
Shiny tag objects.
}
\note{
Full credit to the \code{shinydashboard} package for the original implementation of this function.
}
\references{
Shinydashboard package: \url{https://github.com/rstudio/shinydashboard}
}
\keyword{internal}
