% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_page_filter.R
\name{main_page_filter}
\alias{main_page_filter}
\title{Generate Checkbox Group Input Elements}
\usage{
main_page_filter(spreadsheet)
}
\arguments{
\item{spreadsheet}{A data frame containing information about the checkbox groups to be created. The data frame should include columns such as \code{type}, \code{name}, and \code{parameter_list} that define the checkbox group inputs.}
}
\value{
A \code{shinydashboard::box} object containing the generated checkbox group input elements and an action button for submission.
}
\description{
This function generates a set of checkbox group input elements based on the provided checkbox data, and returns them within a Shiny dashboard box along with a submit action button.
}
