% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_page_group.R
\name{main_page_group}
\alias{main_page_group}
\title{Main Page Group}
\usage{
main_page_group(spreadsheet)
}
\arguments{
\item{spreadsheet}{A data frame containing parameters for generating the group input.}
}
\value{
A shiny input widget for grouping data.
}
\description{
This function generates a group input widget for a shinydashboard main page.
}
\details{
This function is designed to create a group input widget for a shinydashboard main page.
It currently supports only one group by.
}
\examples{
# Sample usage
#main_page_group(my_spreadsheet_data)

}
