% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_items.R
\name{menu_items}
\alias{menu_items}
\title{Generate Shiny Dashboard Menu Items}
\usage{
menu_items(contents_list = data_list$contents)
}
\arguments{
\item{contents_list}{A data frame containing the details for each menu item to be generated. The data frame should include columns such as \code{name}, \code{ID}, and \code{icon}.
Default is \code{data_list$contents}.}
}
\value{
A list of \code{shinydashboard} menu items, which can be directly added to a dashboard's sidebar.
}
\description{
This function generates a list of menu items for a \code{shinydashboard} based on the provided contents list. Each menu item corresponds to a tab in the dashboard.
}
