% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_plot_output.R
\name{multiple_plot_output}
\alias{multiple_plot_output}
\title{Return a list of ggplot2 plots}
\usage{
multiple_plot_output(
  data = plhdata_org_clean,
  columns_to_summarise,
  replace = "rp.contact.field.",
  replace_after = NULL,
  plot_type = c("histogram", "boxplot")
)
}
\arguments{
\item{data}{Data frame to calculate summaries from.}

\item{columns_to_summarise}{Variables to summarise.}

\item{replace}{String of which values to remove before the final string.}

\item{replace_after}{String of which values to remove after the final string.}

\item{plot_type}{Whether to display a \code{histogram} or \code{boxplot}}
}
\value{
List of ggplot2 objects.
}
\description{
Return a list of ggplot2 plots
}
\examples{
# TODO
}
