% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_table_output.R
\name{multiple_table_output}
\alias{multiple_table_output}
\title{Generate Multiple Summary Tables}
\usage{
multiple_table_output(
  data = plhdata_org_clean,
  columns_to_summarise,
  replace = "rp.contact.field.",
  replace_after = NULL,
  summaries = c("frequencies", "mean"),
  na.rm = TRUE
)
}
\arguments{
\item{data}{Data frame to calculate summaries from. Default is \code{plhdata_org_clean}.}

\item{columns_to_summarise}{A vector of variable names (as strings) to summarise.}

\item{replace}{A string indicating the pattern to remove from the variable names before displaying them. Default is \code{"rp.contact.field."}.}

\item{replace_after}{A string indicating the pattern to remove after the variable names. Default is \code{NULL}.}

\item{summaries}{A string indicating whether to calculate \code{"frequencies"} or \code{"mean"} summaries. Default is \code{c("frequencies", "mean")}.}

\item{na.rm}{Logical indicating whether to remove \code{NA} values when calculating means. Default is \code{TRUE}.}
}
\value{
A list of summary tables, where each table corresponds to the summary of the respective variable.
}
\description{
This function generates a list of summary tables based on the provided data and columns to summarise.
It supports calculating either frequencies or means for each specified variable.
}
