% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming_conventions.R
\name{naming_conventions}
\alias{naming_conventions}
\title{Standardise Naming Conventions for Variables}
\usage{
naming_conventions(x, replace, replace_after)
}
\arguments{
\item{x}{A vector of strings containing variable names to standardise.}

\item{replace}{A string indicating the pattern to remove from the start of the variable names.}

\item{replace_after}{A string indicating the pattern to remove from the end of the variable names.}
}
\value{
A vector of strings containing the standardized variable names.
}
\description{
This function standardises the naming conventions for a given set of variable names by removing specified patterns and capitalising the first letter.
}
\examples{
# Example usage:
data_hp_started <- c("rp.contact.field.w_1on1_hp_review_started",
                     "rp.contact.field.w_praise_hp_review_started",
                     "rp.contact.field.w_instruct_hp_review_started",
                     "rp.contact.field.w_stress_hp_review_started")
naming_conventions(data_hp_started, replace = "rp.contact.field.w_",
                   replace_after = "_hp_review_started")
}
