% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLH_shiny.R
\name{PLH_shiny}
\alias{PLH_shiny}
\title{PLH SHINY FUNCTION - exploring using observe() and lazy load.}
\usage{
PLH_shiny(
  title,
  data_list,
  data_frame,
  status = "primary",
  colour = "blue",
  date_from = "2021-10-14",
  key_var = NULL
)
}
\arguments{
\item{title}{Title of the dashboard.}

\item{data_frame}{Spreadsheet that contains information to put in the box.}

\item{colour}{Skin colour of the Shiny App.}

\item{date_from}{Initial date to filter from.}

\item{key_var}{Name of the variable to link different data frames together.}

\item{spreadsheet}{Spreadsheet that contains meta information to put in the box.}
}
\value{
Shiny App
}
\description{
Function to create Shiny
}
