% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_box_function.R
\name{server_box_function}
\alias{server_box_function}
\title{Server Box Function}
\usage{
server_box_function(data_frame, spreadsheet, unique_ID)
}
\arguments{
\item{data_frame}{The data frame used for generating box plots and frequency tables.}

\item{spreadsheet}{A data frame containing spreadsheet data that defines the operations and filters to be applied.}

\item{unique_ID}{An identifier for the current operation.}
}
\value{
A list containing table and plot objects resulting from the processing of the spreadsheet data.
}
\description{
This function processes and generates box plots and frequency tables for a Shiny application server. It reads data from a data frame and applies specified filters and operations based on spreadsheet data.
}
\examples{
# Process spreadsheet data using server_box_function
data_frame <- data.frame(...)  # Define your data frame
spreadsheet_data <- data.frame(...)  # Define your spreadsheet data
unique_ID <- "boxplot_variable1"
result <- server_box_function(data_frame, spreadsheet_data, unique_ID)

# Access table and plot objects from the result
table_obj <- result$table_obj
plot_obj <- result$plot_obj
}
