% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_display_contents.R
\name{server_display_contents}
\alias{server_display_contents}
\title{Server Display Contents}
\usage{
server_display_contents(
  contents1 = contents,
  data_frame,
  data_list,
  loop = NULL,
  list_of_reactives,
  id_name,
  k = 1
)
}
\arguments{
\item{contents1}{A list or data frame containing metadata and information about the content to be displayed.}

\item{data_frame}{The primary data frame used to generate the displayed content.}

\item{data_list}{A list of additional data frames used in various parts of the display.}

\item{loop}{Optional. A vector of loop indices for nested tabbed displays.}

\item{list_of_reactives}{A list of reactive expressions used to update content dynamically in Shiny.}

\item{id_name}{A string specifying the ID of the content to be displayed.}

\item{k}{Optional. An integer index for looping through tabbed displays.}
}
\value{
A list of display boxes that can be rendered in a Shiny UI.
}
\description{
This function processes and displays content within a Shiny application, including handling display sheets and tabbed displays. It organises data into a structured format suitable for visualisation.
}
