% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_display_contents_tabbed.R
\name{server_display_contents_tabbed}
\alias{server_display_contents_tabbed}
\title{Server Display Contents}
\usage{
server_display_contents_tabbed(
  contents1 = contents,
  data_frame,
  data_list,
  loop = NULL,
  list_of_reactives,
  k = 1
)
}
\arguments{
\item{contents1}{A list containing information about the content to be displayed.}

\item{data_frame}{The data frame used for displaying content.}

\item{data_list}{A list containing data associated with the content.}

\item{loop}{A vector of loop indices, used for nested tabbed displays. Default is NULL.}

\item{k}{An integer specifying the initial value for looping through tabbed displays. Default is 1.}
}
\value{
A list of display boxes, each containing the content to be displayed.
}
\description{
This function processes and displays contents, including display and tabbed display sheets, in a Shiny application server. It allows you to organize and present data frames and data lists in a structured way.
}
