% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_calculation.R
\name{summary_calculation}
\alias{summary_calculation}
\title{Calculate Summary Statistics}
\usage{
summary_calculation(
  data = plhdata_org_clean,
  factors,
  columns_to_summarise = NULL,
  summaries = c("frequencies", "mean"),
  together = FALSE,
  include_margins = FALSE,
  drop = FALSE
)
}
\arguments{
\item{data}{A data frame to calculate summaries from.}

\item{factors}{A list of factors to group by.}

\item{columns_to_summarise}{Variables to summarise using \code{dplyr::summarise}.}

\item{summaries}{A character string indicating whether to calculate \code{frequencies} or \code{mean} summaries.}

\item{together}{Logical. Default \code{FALSE}. If \code{summaries = "frequencies"}, whether to combine the count and percentage into one cell.}

\item{include_margins}{Logical. Default \code{FALSE}. Whether to include margins in the summary.}

\item{drop}{Logical. Default \code{FALSE}. Whether to drop unused levels in grouping factors.}
}
\value{
A summary table as a \code{tibble}.
}
\description{
This function calculates either frequency counts or mean summaries from a given dataset.
It supports grouping by factors and can include marginal summaries if desired.
}
