% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot.R
\name{summary_plot}
\alias{summary_plot}
\title{Plot Summaries from PLH Data}
\usage{
summary_plot(
  data = plhdata_org_clean,
  columns_to_summarise,
  naming_convention = TRUE,
  replace = "rp.contact.field.",
  replace_after = NULL,
  plot_type = c("histogram", "boxplot")
)
}
\arguments{
\item{data}{A data frame to plot.}

\item{columns_to_summarise}{The variable to plot.}

\item{naming_convention}{Logical. Default \code{TRUE}. Whether to apply naming conventions to axis labels.}

\item{replace}{A string indicating values to remove from the \code{columns_to_summarise} variable (before the value to keep).}

\item{replace_after}{A string indicating values to remove from the \code{columns_to_summarise} variable (after the value to keep).}

\item{plot_type}{A character string indicating whether to create a \code{histogram} or \code{boxplot}.}
}
\value{
A \code{ggplot} object representing the plot.
}
\description{
This function generates either a histogram or boxplot for a specified variable in PLH data.
It allows for customisation of axis labels and supports naming conventions.
}
