% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Calculate and format summaries from PLH data}
\usage{
summary_table(
  data = plhdata_org_clean,
  factors = Org,
  columns_to_summarise = NULL,
  summaries = c("frequencies", "mean"),
  replace = "rp.contact.field.",
  include_margins = FALSE,
  wider_table = TRUE,
  display_table = FALSE,
  naming_convention = TRUE,
  include_percentages = FALSE,
  together = TRUE,
  drop = FALSE
)
}
\arguments{
\item{data}{Data frame to calculate summaries from.}

\item{factors}{List of factors to group by.}

\item{columns_to_summarise}{Variables to summarise.}

\item{summaries}{Whether \code{frequencies} or \code{mean} summaries are calculated.}

\item{replace}{String of values in the \code{columns_to_summarise} variable to remove in the table (before the value to keep).}

\item{include_margins}{logical. Default \code{FALSE}. Whether to include margins.}

\item{wider_table}{logical. Default \code{TRUE}. Whether to have a wider table if \code{summaries = "frequencies"}.}

\item{display_table}{logical. Default \code{FALSE}. Return table in \code{gt::gt()} table format.}

\item{naming_convention}{logical. Default \code{TRUE}. Whether to apply naming conventions when \code{summaries = "mean"}.}

\item{include_percentages}{logical. Default \code{FALSE}. Whether to include percentages when \code{summaries = "frequencies"}.}

\item{together}{logical. Default \code{FALSE}. If \code{summaries = "frequencies"}, whether to combine the count and percentage into one cell.}

\item{drop}{logical. Default \code{FALSE}. When running \code{group_by}, whether to drop unused columns.}
}
\value{
Summaries table as \code{tibble} or \code{gt}.
}
\description{
Calculate and format summaries table
}
\examples{
# TODO
}
