% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_assert.R
\name{tag_assert}
\alias{tag_assert}
\title{Assert that a tag has specified properties}
\usage{
tag_assert(tag, type = NULL, class = NULL, allowUI = TRUE)
}
\arguments{
\item{tag}{A tag object (e.g., a \code{div} or \code{span} created with Shiny functions).}

\item{type}{The expected type of the tag, like "div", "a", or "span". If \code{NULL}, the type is not checked.}

\item{class}{An expected HTML class. If \code{NULL}, the class is not checked.}

\item{allowUI}{If \code{TRUE} (the default), allows dynamic outputs generated by
\code{\link[shiny]{uiOutput}} or \code{\link[shiny]{htmlOutput}}. When a
dynamic output is provided, \code{tagAssert} won't try to validate the contents.}
}
\description{
This function is a version of the \code{tagAssert} function
from the \code{shinydashboard} package. It checks if a tag has the
specified properties, such as type and class.
}
\details{
This function is from by the internal \code{tagAssert} function
in the \code{shinydashboard} package, which performs similar validations on
Shiny tag objects.
}
\note{
Full credit to the \code{shinydashboard} package for the original implementation of this function.
}
\references{
Shinydashboard package: \url{https://github.com/rstudio/shinydashboard}
}
