% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_na_variable.R
\name{add_na_variable}
\alias{add_na_variable}
\title{Add a variable to the data containing NA values}
\usage{
add_na_variable(data, variable)
}
\arguments{
\item{data}{Data frame to add the variable to.}

\item{variable}{Variable name to add in the data frame.}
}
\value{
Data frame
}
\description{
If a variable is missing in the data frame, add in that variable (with a warning) with NA as its values
}
\examples{
# TODO
}
