% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_na_variable.R
\name{add_na_variable}
\alias{add_na_variable}
\title{Add NA Values to Missing Variables}
\usage{
add_na_variable(data, variable)
}
\arguments{
\item{data}{A data frame to modify.}

\item{variable}{A character vector specifying the names of the variables to check
and potentially add with \code{NA} values.}
}
\value{
The modified data frame with the specified variables added if they were missing,
containing only \code{NA} values.
}
\description{
This function checks if the specified variables exist in a given data frame.
If any of the specified variables do not exist, they are added to the data frame
with all values set to \code{NA}.
}
\examples{
# Add missing variables to the 'mtcars' data frame
df <- add_na_variable(data = mtcars, variable = c("cyl", "var2"))

}
