% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_function.R
\name{box_function}
\alias{box_function}
\title{Creating box to be used in \code{PLH_shiny} function}
\usage{
box_function(data_frame, spreadsheet, unique_ID, label_table, label_plot)
}
\arguments{
\item{data_frame}{Data frame that contains the data to analyse.}

\item{spreadsheet}{Spreadsheet that contains the template.}

\item{unique_ID}{Unique identifier.}
}
\value{
Box for use in \code{Shiny}
}
\description{
Creating box to be used in \code{PLH_shiny} function
}
