% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_table.R
\name{boxplot_table}
\alias{boxplot_table}
\title{Creating a Boxplot to be Used in the \code{PLH_shiny} Function}
\usage{
boxplot_table(
  data,
  variable,
  type = c("summary", "freq"),
  spreadsheet,
  grouped_vars = NULL
)
}
\arguments{
\item{data}{Data frame that contains the data to analyse.}

\item{variable}{Character string specifying the variable of interest within the data.}

\item{type}{Character vector specifying the type of table to generate. Options are \code{"summary"} for summary statistics (median, SD) or \code{"freq"} for frequency tables. Default is \code{"summary"}.}

\item{spreadsheet}{List containing commands and templates for data manipulation and plotting.}

\item{grouped_vars}{Optional character string specifying the variable to group by in the table and plot.}
}
\value{
A list containing a summary table and a ggplot boxplot object for use in \code{Shiny}.
}
\description{
This function generates a summary table and a boxplot for use in \code{Shiny} applications, specifically within the \code{PLH_shiny} function. It processes data according to the specifications in the provided spreadsheet.
}
