% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_table.R
\name{boxplot_table}
\alias{boxplot_table}
\title{Creating boxplot to be used in \code{PLH_shiny} function}
\usage{
boxplot_table(
  data,
  variable,
  type = c("summary", "freq"),
  spreadsheet,
  grouped_vars = NULL
)
}
\arguments{
\item{spreadsheet}{Spreadsheet that contains the template.}

\item{data_frame}{Data frame that contains the data to analyse.}

\item{unique_ID}{Unique identifier.}

\item{label_table}{ID for the table.}

\item{label_plot}{ID for the plot.}
}
\value{
Box for use in \code{Shiny}
}
\description{
Creating boxplot to be used in \code{PLH_shiny} function
}
