% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unknown_variables.R
\name{check_unknown_variables}
\alias{check_unknown_variables}
\title{Check Unknown Variables}
\usage{
check_unknown_variables(df)
}
\arguments{
\item{df}{A data frame containing information about variables and data frames.}
}
\value{
A message indicating any unknown variables found.
}
\description{
Check for unknown variables in the provided data frame.
}
\examples{
check_unknown_variables(df = my_df)

}
