% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_contents.R
\name{display_contents}
\alias{display_contents}
\title{Title}
\usage{
display_contents(contents1 = contents, data_frame, loop = NULL, k = 1)
}
\arguments{
\item{contents1}{}

\item{data_frame}{}

\item{loop}{}

\item{k}{}
}
\description{
Title
}
