% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_sheet.R
\name{display_sheet}
\alias{display_sheet}
\title{Title}
\usage{
display_sheet(
  data_list,
  spreadsheet_name,
  d_box,
  status = "primary",
  colour = "blue",
  j = 1
)
}
\arguments{
\item{data_list}{todo}

\item{spreadsheet_name}{todo}

\item{d_box}{todo}

\item{status}{todo}

\item{colour}{todo}

\item{j}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
#todo
}
