% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shet.R
\name{download_sheet}
\alias{download_sheet}
\title{Title}
\usage{
download_sheet(data_list, spreadsheet_name, j = 1)
}
\arguments{
\item{data_list}{todo}

\item{spreadsheet_name}{todo}

\item{j}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
 #todo
}
