% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_page_filter.R
\name{main_page_filter}
\alias{main_page_filter}
\title{Checkbox Group Input}
\usage{
main_page_filter(spreadsheet)
}
\arguments{
\item{checkbox_data}{A data frame containing information about the checkbox groups to be created.}
}
\value{
A list containing checkbox group input elements and an action button.
}
\description{
This function generates a set of checkbox group input elements based on the provided checkbox data.
}
\examples{
## Not Run 
checkbox_group_data2 <- data.frame(type = c("checkbox_group", "checkbox_group"),
                                   name = c("variable", "variable2"),
                                   parameter_list = c('label = "variable", choices = c("Cylinders" = "cyl", "Transmission" = "am", "Gears" = "gear")',
                                                      'label = "variable2", choices = c("Cyls" = "Sepal.Length", "Gers" = "Sepal.Width")'))
ui <- shiny::fluidPage(
 main_page_filter(checkbox_group_data2),
   shiny::tableOutput("data"),
   shiny::tableOutput("data2")
 )
server <- function(input, output, session) {
 output$data <- renderTable({
 mtcars[, c("mpg", input$variable), drop = FALSE]
 }, rownames = TRUE)
 output$data2 <- renderTable({
 iris[, c("Species", input$variable2), drop = FALSE]
 }, rownames = TRUE)
}

#NOT RUN:
#shiny::shinyApp(ui, server)
}
