% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_plot_output.R
\name{multiple_plot_output}
\alias{multiple_plot_output}
\title{Generate Multiple ggplot2 Plots}
\usage{
multiple_plot_output(
  data = plhdata_org_clean,
  columns_to_summarise,
  replace = "rp.contact.field.",
  replace_after = NULL,
  plot_type = c("histogram", "boxplot")
)
}
\arguments{
\item{data}{Data frame to calculate summaries from. Default is \code{plhdata_org_clean}.}

\item{columns_to_summarise}{A vector of variable names (as strings) to summarize and plot.}

\item{replace}{A string indicating the pattern to remove from the variable names before displaying them. Default is \code{"rp.contact.field."}.}

\item{replace_after}{A string indicating the pattern to remove after the variable names. Default is \code{NULL}.}

\item{plot_type}{A string indicating the type of plot to generate, either \code{"histogram"} or \code{"boxplot"}. Default is \code{c("histogram", "boxplot")}.}
}
\value{
A list of ggplot2 objects, where each object is a plot for the corresponding variable.
}
\description{
This function generates a list of ggplot2 plots based on the provided data and columns to summarize.
It supports creating either histograms or boxplots for each specified variable.
}
