% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_table_output.R
\name{multiple_table_output}
\alias{multiple_table_output}
\title{Return a list of summary tables}
\usage{
multiple_table_output(
  data = plhdata_org_clean,
  columns_to_summarise,
  replace = "rp.contact.field.",
  replace_after = NULL,
  summaries = c("frequencies", "mean"),
  na.rm = TRUE
)
}
\arguments{
\item{data}{Data frame to calculate summaries from.}

\item{columns_to_summarise}{Variables to summarise.}

\item{replace}{String of which values to remove before the final string.}

\item{replace_after}{String of which values to remove after the final string.}

\item{summaries}{Whether \code{frequencies} or \code{mean} summaries are calculated.}

\item{na.rm}{logical. Default \code{TRUE}. If \code{summaries = "mean"}, whether to include \code{NA} values.}
}
\value{
List of summary tables.
}
\description{
Return a list of summary tables
}
\examples{
#TODO
}
