% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_table.R
\name{scatter_table}
\alias{scatter_table}
\title{Creating box to be used in \code{PLH_shiny} function}
\usage{
scatter_table(data, variable, type = c("freq", "summary"), spreadsheet)
}
\arguments{
\item{data}{}

\item{variable}{}

\item{type}{}

\item{spreadsheet}{}
}
\value{
table for use in \code{Shiny}
}
\description{
Creating box to be used in \code{PLH_shiny} function
}
