% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_box_function.R
\name{server_box_function}
\alias{server_box_function}
\title{Server Box Function}
\usage{
server_box_function(data_frame, spreadsheet, unique_ID, list_of_reactives)
}
\arguments{
\item{data_frame}{The data frame containing the data to be visualised or summarised.}

\item{spreadsheet}{A data frame specifying the operations, filters, and configurations to be applied.}

\item{unique_ID}{A string that uniquely identifies the operation or plot type being executed.}

\item{list_of_reactives}{A list of reactive data frames used to dynamically update content in a Shiny app.}
}
\value{
A list containing \code{table_obj}, a data frame or table of results, and \code{plot_obj}, a ggplot2 object representing the plot.
}
\description{
This function processes and generates tables and plots based on specified parameters. It is used within a Shiny application to dynamically create visual and tabular content.
}
