% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_display_sheet_setup.R
\name{server_display_sheet_setup}
\alias{server_display_sheet_setup}
\title{Server Display Sheet Setup}
\usage{
server_display_sheet_setup(
  spreadsheet_data,
  data_frame,
  j,
  loop,
  list_of_reactives
)
}
\arguments{
\item{spreadsheet_data}{A data frame containing the metadata and instructions for creating the display.}

\item{data_frame}{The main data frame from which the content will be generated.}

\item{j}{An integer or index for identifying the specific display sheet.}

\item{loop}{Optional. A vector of loop indices for handling nested displays.}

\item{list_of_reactives}{A list of reactive expressions used to dynamically update content in Shiny.}
}
\value{
A list of display boxes for the specified content type.
}
\description{
This function sets up the content and layout for display sheets within a Shiny server. It processes the spreadsheet data and prepares it for visualisation in a Shiny app.
}
