% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_plot.R
\name{specify_plot}
\alias{specify_plot}
\title{Creating box to be used in \code{PLH_shiny} function}
\usage{
specify_plot(data, spreadsheet, grouped_vars = NULL)
}
\value{
Box for use in \code{Shiny}
}
\description{
Creating box to be used in \code{PLH_shiny} function
}
