% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_layout.R
\name{split_layout}
\alias{split_layout}
\title{Title}
\usage{
split_layout(..., cellWidths = NULL, cellArgs = list())
}
\arguments{
\item{...}{todo}

\item{cellWidths}{todo}

\item{cellArgs}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
#todo
}
