% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet_finder.R
\name{spreadsheet_finder}
\alias{spreadsheet_finder}
\title{Find values from a spreadsheet}
\usage{
spreadsheet_finder(data, column)
}
\arguments{
\item{data}{Name of spreadsheet.}

\item{column}{Column to filter to.}
}
\value{
value from column
}
\description{
Find values from a spreadsheet
}
