% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_calculation.R
\name{summary_calculation}
\alias{summary_calculation}
\title{Calculate summaries from PLH data}
\usage{
summary_calculation(
  data = plhdata_org_clean,
  factors,
  columns_to_summarise,
  summaries = c("frequencies", "mean"),
  include_country_margins,
  country_factor,
  together = FALSE,
  include_margins = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{Data frame to calculate summaries from.}

\item{factors}{List of factors to group by.}

\item{columns_to_summarise}{Variables to dplyr::summarise.}

\item{summaries}{Whether \code{frequencies} or \code{mean} summaries are calculated.}

\item{include_country_margins}{logical. Whether to include the country margins.}

\item{country_factor}{If \code{include_country_margins = TRUE}, the factor variable for the country.}

\item{together}{logical. Default \code{FALSE}. If \code{summaries = "frequencies"}, whether to combine the count and percentage into one cell.}

\item{include_margins}{logical. Default \code{FALSE}. Whether to include margins.}

\item{na.rm}{logical. Default \code{TRUE}. If \code{summaries = "mean"}, whether to include \code{NA} values.}
}
\value{
Summaries table
}
\description{
Calculate summaries from PLH data
}
\examples{
# TODO
}
