% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot.R
\name{summary_plot}
\alias{summary_plot}
\title{Plot PLH data}
\usage{
summary_plot(
  data = plhdata_org_clean,
  columns_to_summarise,
  naming_convention = TRUE,
  replace = "rp.contact.field.",
  replace_after = NULL,
  plot_type = c("histogram", "boxplot")
)
}
\arguments{
\item{data}{Data frame to calculate summaries from.}

\item{columns_to_summarise}{Variables to dplyr::summarise.}

\item{naming_convention}{logical. Default \code{TRUE}. Whether to apply naming conventions.}

\item{replace}{String of values in the \code{columns_to_summarise} variable to remove in the table (before the value to keep).}

\item{replace_after}{String of values in the \code{columns_to_summarise} variable to remove in the table (after the value to keep).}

\item{plot_type}{Whether to display a \code{histogram} or \code{boxplot}}
}
\value{
\code{ggplot} object
}
\description{
Plot either a histogram or boxplot for PLH data.
}
\examples{
# TODO
}
