% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Create and Format Summary Tables from PLH Data}
\usage{
summary_table(
  data = plhdata_org_clean,
  factors = Org,
  columns_to_summarise = NULL,
  summaries = c("frequencies", "mean"),
  replace = "rp.contact.field.",
  include_margins = FALSE,
  wider_table = TRUE,
  display_table = FALSE,
  naming_convention = TRUE,
  include_percentages = FALSE,
  together = TRUE,
  drop = FALSE
)
}
\arguments{
\item{data}{A data frame to summarise.}

\item{factors}{A list of factors to group by.}

\item{columns_to_summarise}{Variables to summarise.}

\item{summaries}{A character string indicating whether to calculate \code{frequencies} or \code{mean} summaries.}

\item{replace}{A string of values in the \code{columns_to_summarise} variable to remove in the table (before the value to keep).}

\item{include_margins}{Logical. Default \code{FALSE}. Whether to include margins in the summary.}

\item{wider_table}{Logical. Default \code{TRUE}. Whether to present a wider table if \code{summaries = "frequencies"}.}

\item{display_table}{Logical. Default \code{FALSE}. Whether to return the table in \code{gt::gt()} format.}

\item{naming_convention}{Logical. Default \code{TRUE}. Whether to apply naming conventions when \code{summaries = "mean"}.}

\item{include_percentages}{Logical. Default \code{FALSE}. Whether to include percentages when \code{summaries = "frequencies"}.}

\item{together}{Logical. Default \code{FALSE}. If \code{summaries = "frequencies"}, whether to combine the count and percentage into one cell.}

\item{drop}{Logical. Default \code{FALSE}. When running \code{group_by}, whether to drop unused columns.}
}
\value{
A summary table as a \code{tibble} or a \code{gt} table, depending on the \code{display_table} parameter.
}
\description{
This function calculates and formats summary tables from PLH data, supporting both frequency and mean summaries.
It can output the table in a wide format, apply naming conventions, and optionally display the table using \code{gt}.
}
