% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabbed_sheet.R
\name{tabbed_sheet}
\alias{tabbed_sheet}
\title{Title}
\usage{
tabbed_sheet(
  data_list = data_list,
  spreadsheet_name = data_list$contents$ID[[i]],
  d_box,
  j = 1
)
}
\arguments{
\item{data_list}{todo}

\item{spreadsheet_name}{todo}

\item{d_box}{todo}

\item{j}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
#todo
}
