% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_value_boxes.R
\name{top_value_boxes}
\alias{top_value_boxes}
\title{Creating top value box to be used in \code{PLH_shiny} function}
\usage{
top_value_boxes(data_frame, spreadsheet, processed_spreadsheet, unique_ID)
}
\arguments{
\item{data_frame}{Data frame that contains the data to analyse.}

\item{spreadsheet}{Spreadsheet that contains the template.}

\item{unique_ID}{Unique identifier.}
}
\value{
Top value box for use in \code{Shiny}
}
\description{
Creating top value box to be used in \code{PLH_shiny} function
}
